backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0

# Vector DB
chromadb>=0.5.0

# Document parsing
pymupdf>=1.24.0
python-docx>=1.1.0

# Configuration
pydantic-settings>=2.0.0

# Development
ruff>=0.6.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
